# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: iTodo API
  version: v1
  description: Cross-platform TodoList HTTP API for mini-program, web, and mobile clients.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/todos:
    get:
      tags: [Todos]
      summary: Query todos
      parameters:
        - name: listId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED]
        - name: important
          in: query
          schema:
            type: boolean
        - name: myDay
          in: query
          schema:
            type: boolean
        - name: dueFrom
          in: query
          schema:
            type: string
            format: date
        - name: dueTo
          in: query
          schema:
            type: string
            format: date
        - name: remindFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: remindTo
          in: query
          schema:
            type: string
            format: date-time
        - name: keyword
          in: query
          schema:
            type: string
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          $ref: '#/components/responses/TodoPageSuccess'
    post:
      tags: [Todos]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
...
  /api/v1/tags:
    get:
      tags: [Tags]
      summary: List tags
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListEnvelope'
    post:
      tags: [Tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '200':
          $ref: '#/components/responses/TagSuccess'
  /api/v1/tags/{tagId}:
    patch:
      tags: [Tags]
      summary: Update a tag
      parameters:
        - $ref: '#/components/parameters/TagId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          $ref: '#/components/responses/TagSuccess'
    delete:
      tags: [Tags]
      summary: Delete a tag
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        '200':
          description: Deleted
  /api/v1/todos/{todoId}/tags/{tagId}:
    post:
      tags: [Tags]
      summary: Bind tag to todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
        - $ref: '#/components/parameters/TagId'
      responses:
        '200':
          description: Bound
    delete:
      tags: [Tags]
      summary: Unbind tag from todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
        - $ref: '#/components/parameters/TagId'
      responses:
        '200':
          description: Unbound
  /api/v1/reminders:
    get:
      tags: [Reminders]
      summary: List reminders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, SENT, CANCELLED]
        - name: remindFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: remindTo
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          $ref: '#/components/responses/ReminderPageSuccess'
  /api/v1/todos/{todoId}/reminders:
    post:
      tags: [Reminders]
      summary: Create a reminder for a todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '200':
          $ref: '#/components/responses/ReminderSuccess'
  /api/v1/todos/{todoId}/reminders/{reminderId}:
    patch:
      tags: [Reminders]
      summary: Update a reminder
      parameters:
        - $ref: '#/components/parameters/TodoId'
        - $ref: '#/components/parameters/ReminderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          $ref: '#/components/responses/ReminderSuccess'
    delete:
      tags: [Reminders]
      summary: Delete a reminder
      parameters:
        - $ref: '#/components/parameters/TodoId'
        - $ref: '#/components/parameters/ReminderId'
      responses:
        '200':
          description: Deleted
  /api/v1/search/todos:
    get:
      tags: [Search]
      summary: Search todos by keyword
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          $ref: '#/components/responses/TodoPageSuccess'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TodoId:
      name: todoId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    TagId:
      name: tagId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ReminderId:
      name: reminderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    TodoSuccess:
      description: Todo operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TodoResponseEnvelope'
    TodoPageSuccess:
      description: Todo page query succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TodoPageResponseEnvelope'
...
    TagSuccess:
      description: Tag operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TagResponseEnvelope'
    ReminderSuccess:
      description: Reminder operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReminderResponseEnvelope'
    ReminderPageSuccess:
      description: Reminder page query succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReminderPageResponseEnvelope'
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        listId: { type: string, format: uuid }
        title: { type: string, maxLength: 255 }
        note: { type: string }
        dueDate: { type: string, format: date }
        remindAt: { type: string, format: date-time }
        repeatRule: { type: string, maxLength: 255 }
        myDay: { type: boolean }
        important: { type: boolean }
        sortOrder: { type: integer }
...
    TodoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        listId: { type: string, format: uuid }
        title: { type: string }
        note: { type: string }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        importance: { type: string, enum: [NORMAL, IMPORTANT] }
        dueDate: { type: string, format: date }
        remindAt: { type: string, format: date-time }
        repeatRule: { type: string }
        completedAt: { type: string, format: date-time }
        sortOrder: { type: integer }
        myDay: { type: boolean }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 64 }
        color: { type: string, maxLength: 32 }
    UpdateTagRequest:
      type: object
      properties:
        name: { type: string, maxLength: 64 }
        color: { type: string, maxLength: 32 }
    TagResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateReminderRequest:
      type: object
      required: [remindAt]
      properties:
        remindAt: { type: string, format: date-time }
        channel: { type: string, enum: [IN_APP] }
    UpdateReminderRequest:
      type: object
      properties:
        remindAt: { type: string, format: date-time }
        channel: { type: string, enum: [IN_APP] }
        status: { type: string, enum: [PENDING, SENT, CANCELLED] }
    ReminderResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        todoId: { type: string, format: uuid }
        remindAt: { type: string, format: date-time }
        channel: { type: string }
        status: { type: string, enum: [PENDING, SENT, CANCELLED] }
        sentAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TodoResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/TodoResponse' }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
    TodoPageResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/TodoResponse' }
            total: { type: integer, format: int64 }
            page: { type: integer, format: int64 }
            size: { type: integer, format: int64 }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
...
    TagResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/TagResponse' }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
    TagListEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { $ref: '#/components/schemas/TagResponse' }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
    ReminderResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/ReminderResponse' }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
    ReminderPageResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ReminderResponse' }
            total: { type: integer, format: int64 }
            page: { type: integer, format: int64 }
            size: { type: integer, format: int64 }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
...