# src/main/java/com/example/itodo/reminder/dto/CreateReminderRequest.java
package com.example.itodo.reminder.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

import java.time.Instant;

public record CreateReminderRequest(
        @NotNull Instant remindAt,
        @Size(max = 32) String channel
) {
}